Added client status, polished login, and moved main files to src directory.
