--- bot.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const {Client, MessageAttachment} = require("discord.js");
-const client = new Client();
-
-
-module.exports = {
-  client,
-  MessageAttachment
-}
