+++ events/guildMemberRemove.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const utils = require("../utils.js");
+const utils = require("../src/utils.js");
 
 
 module.exports = {
